requirements.txt
requests
rich
pyttsx3
speechrecognition
pyaudio
wikipedia
pywhatkit
python-dotenv
AppOpener